+++ src/field/graphics/MeshBuilder.java
@@ -233,7 +233,7 @@ public class MeshBuilder implements MeshAcceptor, Bracketable {
 	 * Adds some aux data associated with attribute "attribute" with dimension 2. "attribute" must be between 1 and 15 inclusive
 	 */
 	public MeshBuilder aux(int attribute, Vec2 x) {
-		aux.put(attribute, new float[]{x.x, x.y});
+		aux.put(attribute, new float[]{(float)x.x, (float)x.y});
 		ensureExists(attribute, 2, vertexCursor);
 		return this;
 	}
@@ -242,7 +242,7 @@ public class MeshBuilder implements MeshAcceptor, Bracketable {
 	 * Adds some aux data associated with attribute "attribute" with dimension 3. "attribute" must be between 1 and 15 inclusive
 	 */
 	public MeshBuilder aux(int attribute, Vec3 x) {
-		aux.put(attribute, new float[]{x.x, x.y, x.z});
+		aux.put(attribute, new float[]{(float)x.x, (float)x.y, (float)x.z});
 		ensureExists(attribute, 3, vertexCursor);
 		return this;
 	}
@@ -251,7 +251,7 @@ public class MeshBuilder implements MeshAcceptor, Bracketable {
 	 * Adds some aux data associated with attribute "attribute" with dimension 4. "attribute" must be between 1 and 15 inclusive
 	 */
 	public MeshBuilder aux(int attribute, Vec4 x) {
-		aux.put(attribute, new float[]{x.x, x.y, x.z, x.w});
+		aux.put(attribute, new float[]{(float)x.x, (float)x.y, (float)x.z, (float)x.w});
 		ensureExists(attribute, 4, vertexCursor);
 		return this;
 	}
